feat: include wrap in swap info trade state
